tutorials/llama_index_rag_serving/requirements.txt
adlfs==2024.12.0
aiobotocore==2.21.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.33.0
azure-datalake-store==0.0.53
azure-identity==1.21.0
azure-storage-blob==12.25.1
backports-tarfile==1.2.0
banks==2.1.2
beautifulsoup4==4.13.4
botocore==1.37.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
croniter==6.0.0
cryptography==44.0.2
cssselect==1.3.0
dataclasses-json==0.5.9
decorator==5.2.1
defusedxml==0.7.1
deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring-parser==0.16
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
filetype==1.2.0
flyteidl==1.15.3
flytekit==1.15.4
frozenlist==1.6.0
fsspec==2025.3.2
gcsfs==2025.3.2
google-api-core==2.24.2
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
griffe==1.7.3
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.16.0
html2text==2024.2.26
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib-metadata==8.6.1
isodate==0.7.2
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.1.0
jieba3k==0.35.1
jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonpickle==4.0.5
keyring==25.6.0
lancedb==0.21.2
llama-index-agent-openai==0.4.6
llama-index-core==0.12.33.post1
llama-index-embeddings-openai==0.3.1
llama-index-embeddings-openai-like==0.1.0
llama-index-llms-openai==0.3.38
llama-index-llms-openai-like==0.3.4
llama-index-readers-web==0.3.9
llama-index-vector-stores-lancedb==0.3.2
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
marshmallow-enum==1.5.1
marshmallow-jsonschema==0.13.0
mashumaro==3.15
mdurl==0.1.2
more-itertools==10.7.0
msal==1.32.3
msal-extensions==1.3.1
msgpack==1.1.0
multidict==6.4.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
obstore==0.3.0
openai==1.76.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
protoc-gen-openapiv2==0.0.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pyee==12.1.1
pygments==2.19.1
pyjwt==2.10.1
pylance==0.26.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==14.0.0
rich-click==1.8.8
rich-toolkit==0.14.3
rsa==4.9.1
s3fs==2025.3.2
safetensors==0.5.3
selenium==4.31.0
sentry-sdk==2.27.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spider-client==0.0.27
sqlalchemy==2.0.40
starlette==0.46.2
statsd==4.0.1
tantivy==0.22.2
tenacity==9.1.2
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
union==0.1.180
union-runtime==0.1.17
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
